+++ src/platforms/linux/syscalls/io_syscalls.cpp
@@ -1603,6 +1603,8 @@ uint64_t sys_openat(std::shared_ptr<Arion> arion, std::vector<SYS_PARAM> params)
     int32_t flags = params.at(2);
     uint16_t mode = params.at(3);
 
+    flags &= ~0x20000;
+
     std::string file_name = arion->mem->read_c_string(file_name_addr);
     std::string abs_file_path = get_path_at(arion, dfd, file_name);
     int fd = open(abs_file_path.c_str(), flags, mode);
@@ -2137,8 +2139,14 @@ uint64_t sys_statx(std::shared_ptr<Arion> arion, std::vector<SYS_PARAM> params)
     int statx_ret = statx(0, abs_file_path.c_str(), flags, mask, &statx_buf);
     if (statx_ret == -1)
         statx_ret = -errno;
-    else
-        arion->mem->write(statx_buf_addr, (BYTE *)&statx_buf, sizeof(struct stat));
+    else {
+        STRUCT_ID statx_id = STATX_STRUCT_FACTORY.feed_host(&statx_buf);
+        size_t arch_statx_len;
+        BYTE *arch_statx = STATX_STRUCT_FACTORY.build(statx_id, arion->abi->get_attrs()->arch, arch_statx_len);
+        arion->mem->write(statx_buf_addr, arch_statx, arch_statx_len);
+        free(arch_statx);
+        STATX_STRUCT_FACTORY.release_struct(statx_id);
+    }
     return statx_ret;
 }